tlsKeyStore moved to its own file and renamed
